requirements.txt
# Market Anomaly Detection Engine - Dependencies
# Python 3.9+

# ============================================
# CORE DATA PROCESSING
# ============================================
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# ============================================
# MACHINE LEARNING
# ============================================
scikit-learn==1.3.2

# Deep Learning (for Autoencoder & LSTM)
tensorflow==2.15.0
keras==2.15.0

# ============================================
# FINANCIAL DATA
# ============================================
yfinance==0.2.33

# ============================================
# API & WEB FRAMEWORK
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-multipart==0.0.6

# ============================================
# VISUALIZATION & DASHBOARD
# ============================================
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# ============================================
# EXPLAINABLE AI (XAI)
# ============================================
shap==0.43.0
lime==0.2.0.1

# ============================================
# DATABASE
# ============================================
sqlalchemy==2.0.23

# ============================================
# EXPERIMENT TRACKING & MODEL MANAGEMENT
# ============================================
mlflow==2.9.2
mlflow-skinny==2.9.2

# ============================================
# UTILITIES
# ============================================
python-dotenv==1.0.0
requests==2.31.0

# ============================================
# DEVELOPMENT & TESTING
# ============================================
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1
isort==5.13.2
pre-commit==3.6.0

# Type stubs
types-requests==2.31.0.10
pandas-stubs==2.1.4.231227
